+++ docs/changelog.md
@@ -3,6 +3,7 @@
 * Added `KurtosisContext.GetUserServiceLogs` method in `golang` and `typescript` api libraries
 * Added the public documentation for the new `KurtosisContext.GetUserServiceLogs` method
 * Added `GetUserServiceLogs` in Kurtosis engine gateway
+* Implemented IP address references for services
 
 ### Changes
 * Add back old enclave continuity test